41 Moor End
"Yes, sir. I met him that night, and he called next day to ask if we had got home all safe, and
after that we met him--that is to say, Mr. Holmes, I met him twice for walks, but after that father
came back again, and Mr. Hosmer Angel could not come to the house any more."

"No?"