119 Harvey Street
"Why did you come away to consult me in such a hurry?" asked Sherlock Holmes, with his finger-tips
together and his eyes to the ceiling.

Again a startled look came over the somewhat vacuous face of Miss Mary Sutherland. "Yes, I did bang
out of the house," she said, "for it made me angry to see the easy way in which Mr. Windibank--that
is, my father--took it all. He would not go to the police, and he would not go to you, and so at
last, as he would do nothing and kept on saying that there was no harm done, it made me mad, and I
just on with my things and came right away to you."

"Your father," said Holmes, "your stepfather, surely, since the name is different."

"Yes, my stepfather. I call him father, though it sounds funny, too, for he is only five years and
two months older than myself."

"And your mother is alive?"

"Oh, yes, mother is alive and well. I wasn't best pleased, Mr. Holmes, when she married again so
soon after father's death, and a man who was nearly fifteen years younger than herself. Father was a
plumber in the Tottenham Court Road, and he left a tidy business behind him, which mother carried on
with Mr. Hardy, the foreman; but when Mr. Windibank came he made her sell the business, for he was
very superior, being a traveller in wines. They got £4700 for the goodwill and interest, which
wasn't near as much as father could have got if he had been alive."