63 Down Close
"'Well, you have to be in the office, or at least in the building, the whole time. If you leave, you
forfeit your whole position forever. The will is very clear upon that point. You don't comply with
the conditions if you budge from the office during that time.'

"'It's only four hours a day, and I should not think of leaving,' said I.

"'No excuse will avail,' said Mr. Duncan Ross; 'neither sickness nor business nor anything else.
There you must stay, or you lose your billet.'

"'And the work?'

"'Is to copy out the "Encyclopaedia Britannica." There is the first volume of it in that press. You
must find your own ink, pens, and blotting-paper, but we provide this table and chair. Will you be
ready to-morrow?'