168 Down Close
Holmes had sat up upon the couch, and I saw him motion like a man who is in need of air. A maid
rushed across and threw open the window. At the same instant I saw him raise his hand and at the
signal I tossed my rocket into the room with a cry of "Fire!" The word was no sooner out of my mouth
than the whole crowd of spectators, well dressed and ill--gentlemen, ostlers, and servant-maids--
joined in a general shriek of "Fire!" Thick clouds of smoke curled through the room and out at the
open window. I caught a glimpse of rushing figures, and a moment later the voice of Holmes from
within assuring them that it was a false alarm. Slipping through the shouting crowd I made my way to
the corner of the street, and in ten minutes was rejoiced to find my friend's arm in mine, and to
get away from the scene of uproar. He walked swiftly and in silence for some few minutes until we
had turned down one of the quiet streets which lead towards the Edgeware Road.

"You did it very nicely, Doctor," he remarked. "Nothing could have been better. It is all right."

"You have the photograph?"

"I know where it is."

"And how did you find out?"

"She showed me, as I told you she would."