171 Glan Road
"This may interest you, Lestrade," he remarked, holding it out.  "The murder was done with it."

"I see no marks."

"There are none."

"How do you know, then?"

"The grass was growing under it. It had only lain there a few days. There was no sign of a place
whence it had been taken. It corresponds with the injuries. There is no sign of any other weapon."

"And the murderer?"