8 Tamworth Drive
"'Quite so. It is very natural that the pledge of secrecy which we have exacted from you should have
aroused your curiosity. I have no wish to commit you to anything without your having it all laid
before you. I suppose that we are absolutely safe from eavesdroppers?'

"'Entirely.'

"'Then the matter stands thus. You are probably aware that fuller's-earth is a valuable product, and
that it is only found in one or two places in England?'

"'I have heard so.'