101 Beaconside
"What do you make of that?" asked Holmes.

"The name of the maker, no doubt; or his monogram, rather."

"Not at all. The 'G' with the small 't' stands for 'Gesellschaft,' which is the German for
'Company.' It is a customary contraction like our 'Co.' 'P,' of course, stands for 'Papier.' Now for
the 'Eg.' Let us glance at our Continental Gazetteer." He took down a heavy brown volume from his
shelves. "Eglow, Eglonitz--here we are, Egria. It is in a German-speaking country--in Bohemia, not
far from Carlsbad. 'Remarkable as being the scene of the death of Wallenstein, and for its numerous
glass-factories and paper-mills.' Ha, ha, my boy, what do you make of that?" His eyes sparkled, and
he sent up a great blue triumphant cloud from his cigarette.

"The paper was made in Bohemia," I said.

"Precisely. And the man who wrote the note is a German. Do you note the peculiar construction of the
sentence--'This account of you we have from all quarters received.' A Frenchman or Russian could not
have written that. It is the German who is so uncourteous to his verbs. It only remains, therefore,
to discover what is wanted by this German who writes upon Bohemian paper and prefers wearing a mask
to showing his face. And here he comes, if I am not mistaken, to resolve all our doubts."