186 Glan Road
"I see," said I as I glanced down the column, "that the coroner in his concluding remarks was rather
severe upon young McCarthy. He calls attention, and with reason, to the discrepancy about his father
having signalled to him before seeing him, also to his refusal to give details of his conversation
with his father, and his singular account of his father's dying words. They are all, as he remarks,
very much against the son."

Holmes laughed softly to himself and stretched himself out upon the cushioned seat. "Both you and
the coroner have been at some pains," said he, "to single out the very strongest points in the young
man's favour. Don't you see that you alternately give him credit for having too much imagination and
too little? Too little, if he could not invent a cause of quarrel which would give him the sympathy
of the jury; too much, if he evolved from his own inner consciousness anything so outré as a dying
reference to a rat, and the incident of the vanishing cloth. No, sir, I shall approach this case
from the point of view that what this young man says is true, and we shall see whither that
hypothesis will lead us. And now here is my pocket Petrarch, and not another word shall I say of
this case until we are on the scene of action. We lunch at Swindon, and I see that we shall be there
in twenty minutes."