77 Cliff Place
"You will observe," said Holmes, laying down the volume, "that the sudden breaking up of the society
was coincident with the disappearance of Openshaw from America with their papers. It may well have
been cause and effect. It is no wonder that he and his family have some of the more implacable
spirits upon their track. You can understand that this register and diary may implicate some of the
first men in the South, and that there may be many who will not sleep easy at night until it is
recovered."

"Then the page we have seen--"

"Is such as we might expect. It ran, if I remember right, 'sent the pips to A, B, and C'--that is,
sent the society's warning to them. Then there are successive entries that A and B cleared, or left
the country, and finally that C was visited, with, I fear, a sinister result for C. Well, I think,
Doctor, that we may let some light into this dark place, and I believe that the only chance young
Openshaw has in the meantime is to do what I have told him. There is nothing more to be said or to
be done to-night, so hand me over my violin and let us try to forget for half an hour the miserable
weather and the still more miserable ways of our fellow-men."