138 Beaconside
"I come to a part of my story now in which I should wish to be particularly so. I am not a very
heavy sleeper, and the anxiety in my mind tended, no doubt, to make me even less so than usual.
About two in the morning, then, I was awakened by some sound in the house. It had ceased ere I was
wide awake, but it had left an impression behind it as though a window had gently closed somewhere.
I lay listening with all my ears. Suddenly, to my horror, there was a distinct sound of footsteps
moving softly in the next room. I slipped out of bed, all palpitating with fear, and peeped round
the corner of my dressing-room door.

"'Arthur!' I screamed, 'you villain! you thief! How dare you touch that coronet?'

"The gas was half up, as I had left it, and my unhappy boy, dressed only in his shirt and trousers,
was standing beside the light, holding the coronet in his hands. He appeared to be wrenching at it,
or bending it with all his strength. At my cry he dropped it from his grasp and turned as pale as
death. I snatched it up and examined it. One of the gold corners, with three of the beryls in it,
was missing.