149 Sunningdale Drive
"None at all," said Holmes. "They are coiners on a large scale, and have used the machine to form
the amalgam which has taken the place of silver."

"We have known for some time that a clever gang was at work," said the inspector. "They have been
turning out half-crowns by the thousand. We even traced them as far as Reading, but could get no
farther, for they had covered their traces in a way that showed that they were very old hands. But
now, thanks to this lucky chance, I think that we have got them right enough."

But the inspector was mistaken, for those criminals were not destined to fall into the hands of
justice. As we rolled into Eyford Station we saw a gigantic column of smoke which streamed up from
behind a small clump of trees in the neighbourhood and hung like an immense ostrich feather over the
landscape.