41 Wantage Close
"'In the case of another,' said he, 'the objection might be fatal, but we must stretch a point in
favour of a man with such a head of hair as yours. When shall you be able to enter upon your new
duties?'

"'Well, it is a little awkward, for I have a business already,' said I.

"'Oh, never mind about that, Mr. Wilson!' said Vincent Spaulding. 'I should be able to look after
that for you.'

"'What would be the hours?' I asked.