87 Cliff Place
"At eight in the morning. She will not be up, so that we shall have a clear field. Besides, we must
be prompt, for this marriage may mean a complete change in her life and habits. I must wire to the
King without delay."

We had reached Baker Street and had stopped at the door. He was searching his pockets for the key
when someone passing said:

"Good-night, Mister Sherlock Holmes."

There were several people on the pavement at the time, but the greeting appeared to come from a slim
youth in an ulster who had hurried by.

"I've heard that voice before," said Holmes, staring down the dimly lit street. "Now, I wonder who
the deuce that could have been."

I slept at Baker Street that night, and we were engaged upon our toast and coffee in the morning
when the King of Bohemia rushed into the room.