35 Badgers Dene
"'What are you doing there?' he asked.

"I felt angry at having been tricked by so elaborate a story as that which he had told me. 'I was
admiring your fuller's-earth,' said I; 'I think that I should be better able to advise you as to
your machine if I knew what the exact purpose was for which it was used.'

"The instant that I uttered the words I regretted the rashness of my speech. His face set hard, and
a baleful light sprang up in his grey eyes.

"'Very well,' said he, 'you shall know all about the machine.' He took a step backward, slammed the
little door, and turned the key in the lock. I rushed towards it and pulled at the handle, but it
was quite secure, and did not give in the least to my kicks and shoves. 'Hullo!' I yelled. 'Hullo!
Colonel! Let me out!'