100 Bowler Street
"Do not lose an instant. And, above all, take care of yourself in the meanwhile, for I do not think
that there can be a doubt that you are threatened by a very real and imminent danger. How do you go
back?"

"By train from Waterloo."

"It is not yet nine. The streets will be crowded, so I trust that you may be in safety. And yet you
cannot guard yourself too closely."

"I am armed."

"That is well. To-morrow I shall set to work upon your case."